refactor: move some code from utils to style
