+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "winshell (>=0.6,<0.7)",
     "win32mica>=3.0",
     "qt-themes>=0.4.0",
+    "winaccent>=2.1.0",
 ]
 
 [dependency-groups]
