+++ src/utility/style.py
@@ -1,12 +1,68 @@
 "Contain styling code"
 
-import qt_themes
+import sys
+import winreg
 from PySide6.QtGui import QPalette, QColor  # pylint: disable=E0611
+from PySide6.QtCore import QRect, Qt  # pylint: disable=E0611
+import qt_themes
+import win32mica
 
 from utility import utils
 
+# ---------------------------- Utility ------------------------------
+def get_geometry(parent_window, width, height):
+    "Get x and y centered relative to parent window"
+    parent_geometry = parent_window.geometry()
+    parent_x = parent_geometry.x()
+    parent_y = parent_geometry.y()
+    parent_width = parent_geometry.width()
+    parent_height = parent_geometry.height()
+
+    x = parent_x + (parent_width - width) // 2
+    y = parent_y + (parent_height - height) // 2
+    return QRect(x, y, width, height)
+
+mica_supported = bool(sys.getwindowsversion().build >= 22000)
+
+def get_theme():
+    "Add system to the theme"
+    current_theme = utils.get_config().theme
+    if current_theme == "system":
+        theme = detect_system_theme()
+    else:
+        theme = current_theme
+
+    return theme
+
+def detect_system_theme():
+    "Detecting system theme for Pyside6 default theme handling"
+    if sys.platform == "win32":
+        try:
+            registry = winreg.ConnectRegistry(None, winreg.HKEY_CURRENT_USER)
+            theme_registry = r"Software\Microsoft\Windows\CurrentVersion\Themes\Personalize"
+            key = winreg.OpenKey(registry, theme_registry)
+            value, _ = winreg.QueryValueEx(key, "AppsUseLightTheme")
+            winreg.CloseKey(key)
+            return "dark" if value == 0 else "light"
+        except FileNotFoundError:
+            print("Theme registry not found.")
+    return "light"
+
+def apply_mica(target_window):
+    "Apply mica style on target window using win32mica"
+    theme = get_theme().upper()
+    mica_effect = utils.get_config().mica_effect.upper()
+
+    if mica_effect != "DISABLE" and mica_supported:
+        target_window.setAttribute(Qt.WA_TranslucentBackground)
+        win32mica.ApplyMica(
+            HWND=int(target_window.winId()),
+            Theme=getattr(win32mica.MicaTheme, theme),
+            Style=getattr(win32mica.MicaStyle, mica_effect)
+        )
+
 # ---------------------------- Variables ----------------------------
-THEME = utils.get_theme()
+THEME = get_theme()
 
 if THEME == "dark":
     SURFACE0 = "rgba(255, 255, 255, 0.052)"
@@ -17,7 +73,7 @@ elif THEME == "light":
     MANTLE = "rgba(0, 0, 0, 0.08)"
     SUBTEXT0 = "rgba(255,255,255,0.8)"
 else:
-    qt_theme_dict = qt_themes.get_theme(utils.get_theme())
+    qt_theme_dict = qt_themes.get_theme(get_theme())
     SURFACE0 = qt_theme_dict.surface0.name()
     MANTLE = qt_theme_dict.mantle.name()
     SUBTEXT0 = qt_theme_dict.subtext0.name()
